Update keepalived to v2.0.19
There is a bug in v2.0.16 of keepalived which is segfaulting when vsg is configured. The fixing commit: https://github.com/acassen/keepalived/commit/50382b81c7ae8720b97073e3e919de12171b1b45 this resolves issue when we upgrade ingress controller and the downloading of new images takes to long time.